requirements.txt
httpx>=0.27.0,
python-dotenv>=0.17.1,
pytest>=6.2.4,